We got data again...
Somehow the Xbox Elite Series 2 Controller doesn`t react but all other test controllers do... this took me almost the entire day because I thought the code was not working properly but it was just this one controller ...

+++ src/main/java/de/gurkenlabs/litiengine/input/windows/DIDEVICEOBJECTDATA.java
@@ -35,7 +35,7 @@ class DIDEVICEOBJECTDATA {
     data.dwData = (int) VH_dwData.get(segment);
     data.dwTimeStamp = (int) VH_dwTimeStamp.get(segment);
     data.dwSequence = (int) VH_dwSequence.get(segment);
-    data.uAppData = (int) VH_uAppData.get(segment);
+    data.uAppData = (long) VH_uAppData.get(segment);
     return data;
   }
 }
